skill-swap-admin/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
python-dotenv
pyjwt
streamlit
pandas
